Added measurementTimer class which allows for persistant state between measurement series. Display of timer shows tenths of a second

+++ display.h
@@ -5,4 +5,4 @@ void display_main_screen();
 void display_measuring_screen();
 
 void display_debug_info(const char *debug_text);
-void display_time(uint32_t tm);
+void display_time(float tm);
